Re-initialise crossplane state
This is a mostly working crossplane/k0rdent setup which
deploys crossplane as a service in a k0rdent clusterdeployment.
This then allows one to connect to that cluster, and apply specific
XR-Claims, thus creating the necessary resources via crossplane.

The merging of many commits to this point is to allow the removal of
complexity as we attempt to achieve the end goal using terraform
definitions.